TASK #4492 路测专题分析功能增强:更新部分功能代码

+++ src/main/java/com/hgicreate/rno/domain/gsm/GsmDtSample.java
@@ -29,8 +29,8 @@ public class GsmDtSample {
     private Integer ci ;
     private String cell;
     private String cellIndoor;
-    private Integer cellLongitude ;
-    private Integer cellLatitude ;
+    private Double cellLongitude ;
+    private Double cellLatitude ;
     private Integer rxlevsub ;
     private Integer rxqualsub ;
     private Integer distance ;
